Fixed bug where create account form data would take it from the login form
The form data had to be taken from the createAccountForm, since the
names of the inputs of the login and create account form were the same,
the jQuery selector would select the login inputs instead since it came
first. Also added in client side validation that the password is the
same as the retyped password when creating an account.