Cached stock prices on server
The stock prices are now cached on the server. Every time a request
comes in, the server checks if the cache was updated in the last 15
minutes. If it was, return the cached stock price. Otherwise, update
the cache by calling Yahoo's Finance API then return the cached stock
price.