Implemented player's cash
Cash is now stored in the database in the users collection. It is
created at the same time an account is created. It is also looked up
after a user logs in and shown on the page. When a user refreshes the
page, it is also looked up and displayed. Cash is always displayed up
to 2 decimal points.